feat(usecase): delete task

+++ src/usecase/task/deleteTaskUseCase.ts
@@ -1,3 +1,4 @@
+import { NotFoundError } from "@/error/notfound.error";
 import { TaskRepository } from "@/repository/taskRepository";
 
 export class DeleteTaskUseCase {
@@ -6,7 +7,10 @@ export class DeleteTaskUseCase {
   ){}
 
   async execute(taskId: string){
-    const task = await this.taskRepository.
+    const task = await this.taskRepository.getById(taskId)
+    if(!task){
+      throw new NotFoundError("task not found")
+    }
     await this.taskRepository.delete(taskId)
   }
 }
